requirements.txt
# =============================================================================
# Genesis Mind — Dependencies (V8: Zero Pretrained Models)
# =============================================================================
# All models are from scratch. No CLIP, no Whisper, no SBERT, no Ollama.
# No CUDA, no GPU drivers, no cloud APIs, no internet required.
# Pinned to exact versions for reproducibility.
# =============================================================================

# --- Core Neural Framework ---
torch==2.11.0                   # Backend for all from-scratch neural networks
torchaudio==2.11.0              # Mel spectrogram computation (pure signal processing)
numpy==2.4.3                    # Numerical backbone for all sensory data

# --- Perception (Senses) ---
opencv-python==4.13.0.92        # Eyes: webcam capture and image processing
sounddevice==0.5.5              # Ears: microphone audio capture
pillow==12.1.1                  # Image manipulation and format conversion

# --- Memory (Vector Database) ---
chromadb==1.5.5                 # Local vector database, no server needed

# --- Voice ---
# Neural vocoder only — no pretrained TTS (pyttsx3 removed)

# --- Utilities ---
scipy==1.17.1                   # Signal processing for audio
Flask==3.1.3                    # Dashboard web server
flask-cors==6.0.2               # CORS support for dashboard
